requirements.txt
asgiref==3.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.4
django-cors-headers==4.9.0
django-filter==25.2
django-watson==1.6.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0